Move get_sdv_logger to logger.py
