+++ sdv/logging/__init__.py
@@ -1,6 +1,7 @@
 """Module for configuring loggers within the SDV library."""
 
-from sdv.logging.utils import disable_single_table_logger, get_sdv_logger, get_sdv_logger_config
+from sdv.logging.logger import get_sdv_logger
+from sdv.logging.utils import disable_single_table_logger, get_sdv_logger_config
 
 __all__ = (
     'disable_single_table_logger',
